# specs/001-lottery-sandbox/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Level Agent Lottery Sandbox API
  description: |
    RESTful API for Multi-Level Agent Lottery Sandbox System supporting Malaysian and Singapore lottery providers.

    **Key Features:**
    - Unlimited agent hierarchy with configurable commission
    - Weekly limit-based betting system (Monday reset)
    - Multiple service providers (Magnum, Sports Toto, Damacai, Singapore Pools)
    - Game types: 3D, 4D, 5D, 6D
    - Bet types: BIG, SMALL, iBox (MVP)
    - Third-party API integration with manual fallback
    - Role-based access control (ADMIN, MODERATOR, AGENT)

  version: 1.0.0
  contact:
    name: API Support
    email: support@lottery-sandbox.com
servers:
  - url: https://api.lottery-sandbox.com/api/v1
    description: Production server
  - url: https://staging-api.lottery-sandbox.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive access token (15min) and refresh token (7 days)
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
            example:
              username: "agent001"
              password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                expiresIn: 900
                user:
                  id: 5
                  username: "agent001"
                  role: "AGENT"
                  uplineId: 2
                  moderatorId: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponseDto'
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                expiresIn: 900
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutDto'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"

  # ============================================================================
  # SERVICE PROVIDERS
  # ============================================================================
...
  /bets:
    get:
      tags:
        - Bets
      summary: List bets
      description: |
        List bets based on user role:
        - ADMIN: All bets across all moderators
        - MODERATOR: Bets within their organization
        - AGENT: Their bets and downline bets
      operationId: listBets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: agentId
          in: query
          schema:
            type: integer
          description: Filter by agent ID
        - name: providerId
          in: query
          schema:
            type: string
          description: Filter by provider ID
        - name: drawId
          in: query
          schema:
            type: integer
          description: Filter by draw ID
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, WON, LOST, CANCELLED, REFUNDED]
          description: Filter by bet status
        - name: gameType
          in: query
          schema:
            type: string
            enum: [3D, 4D, 5D, 6D]
          description: Filter by game type
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter bets from this date (inclusive)
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter bets until this date (inclusive)
      responses:
        '200':
          description: Bets retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BetDto'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - Bets
      summary: Place bet
      description: Place new lottery bet (deducts from weekly limit)
      operationId: placeBet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBetDto'
            example:
              providerId: "clx1a2b3c4d5e6f7g8h9"
              drawId: 1001
              gameType: "4D"
              betType: "BIG"
              selectedNumbers: "1234"
              betAmount: 10.00
      responses:
        '201':
          description: Bet placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetDto'
              example:
                id: 50001
                agentId: 5
                providerId: "clx1a2b3c4d5e6f7g8h9"
                drawId: 1001
                gameType: "4D"
                betType: "BIG"
                selectedNumbers: "1234"
                betAmount: 10.00
                potentialPayout: 25000.00
                status: "PENDING"
                receiptNumber: "BET-20250118-50001"
                createdAt: "2025-01-18T10:30:00Z"
        '400':
          description: Bet validation failed (insufficient quota, duplicate bet, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficientQuota:
                  value:
                    error: "BAD_REQUEST"
                    message: "Insufficient weekly limit. Available: RM 500.00, Required: RM 1000.00"
                    statusCode: 400
                duplicateBet:
                  value:
                    error: "BAD_REQUEST"
                    message: "Duplicate bet detected for this draw and number combination"
                    statusCode: 400
        '422':
          $ref: '#/components/responses/ValidationError'
  /bets/{betId}:
    get:
      tags:
        - Bets
      summary: Get bet details
      operationId: getBet
      parameters:
        - $ref: '#/components/parameters/BetIdParam'
      responses:
        '200':
          description: Bet retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetDto'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Bets
      summary: Cancel bet
      description: Cancel PENDING bet before draw (refunds weekly limit)
      operationId: cancelBet
      parameters:
        - $ref: '#/components/parameters/BetIdParam'
      responses:
        '200':
          description: Bet cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetDto'
              example:
                id: 50001
                status: "CANCELLED"
                cancelledAt: "2025-01-18T11:00:00Z"
        '400':
          description: Bet cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "BAD_REQUEST"
                message: "Cannot cancel bet. Status is not PENDING or draw has already occurred."
                statusCode: 400
  /bets/receipt/{receiptNumber}:
    get:
      tags:
        - Bets
      summary: Get bet by receipt number
      description: Retrieve bet details using receipt number for verification
      operationId: getBetByReceipt
      parameters:
        - name: receiptNumber
          in: path
          required: true
          schema:
            type: string
          example: "BET-20250118-50001"
      responses:
        '200':
          description: Bet retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetDto'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================================
  # WEEKLY LIMITS
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (expires in 15 minutes)

        Example:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
...
    BetIdParam:
      name: betId
      in: path
      required: true
      schema:
        type: integer
      description: Bet ID
...
  schemas:
...
    LoginDto:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "agent001"
        password:
          type: string
          minLength: 8
          maxLength: 100
          example: "SecurePass123!"
    LoginResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (15min expiry)
        refreshToken:
          type: string
          description: JWT refresh token (7 days expiry)
        expiresIn:
          type: integer
          description: Access token expiry in seconds
          example: 900
        user:
          $ref: '#/components/schemas/UserDto'
    RefreshTokenDto:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Valid refresh token
    RefreshTokenResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: New JWT access token
        expiresIn:
          type: integer
          example: 900
    LogoutDto:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    # ========================================================================
    # USER/AGENT SCHEMAS
    # ========================================================================
    UserDto:
      type: object
      properties:
        id:
          type: integer
          example: 5
        username:
          type: string
          example: "agent003"
        role:
          type: string
          enum: [ADMIN, MODERATOR, AGENT]
          example: "AGENT"
        uplineId:
          type: integer
          nullable: true
          example: 2
        moderatorId:
          type: integer
          nullable: true
          example: 1
        weeklyLimit:
          type: number
          format: decimal
          example: 10000.00
        weeklyUsed:
          type: number
          format: decimal
          example: 2500.00
        commissionRate:
          type: number
          format: decimal
          example: 30.00
          description: Commission percentage (0.00-100.00)
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BetDto:
      type: object
      properties:
        id:
          type: integer
          example: 50001
        agentId:
          type: integer
          example: 5
        providerId:
          type: string
          example: "clx1a2b3c4d5e6f7g8h9"
        drawId:
          type: integer
          example: 1001
        gameType:
          type: string
          enum: [3D, 4D, 5D, 6D]
          example: "4D"
        betType:
          type: string
          enum: [BIG, SMALL, IBOX]
          example: "BIG"
        selectedNumbers:
          type: string
          example: "1234"
        betAmount:
          type: number
          format: decimal
          example: 10.00
        potentialPayout:
          type: number
          format: decimal
          example: 25000.00
        actualPayout:
          type: number
          format: decimal
          nullable: true
          example: null
        status:
          type: string
          enum: [PENDING, WON, LOST, CANCELLED, REFUNDED]
          example: "PENDING"
        receiptNumber:
          type: string
          example: "BET-20250118-50001"
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
    CreateBetDto:
      type: object
      required:
        - providerId
        - drawId
        - gameType
        - betType
        - selectedNumbers
        - betAmount
      properties:
        providerId:
          type: string
        drawId:
          type: integer
        gameType:
          type: string
          enum: [3D, 4D, 5D, 6D]
        betType:
          type: string
          enum: [BIG, SMALL, IBOX]
        selectedNumbers:
          type: string
          pattern: '^\d{3,6}$'
          description: 3-6 digit number based on game type
        betAmount:
          type: number
          format: decimal
          minimum: 1
          maximum: 100000

    # ========================================================================
    # WEEKLY LIMIT SCHEMAS
    # ========================================================================
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Insufficient weekly limit"
        statusCode:
          type: integer
          example: 400
        details:
          type: object
          nullable: true
  responses:
    UnauthorizedError:
      description: Authentication failed or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Invalid or expired token"
            statusCode: 401
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"
            statusCode: 404
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "Request validation failed"
            statusCode: 422
            details:
              - field: "betAmount"
                message: "betAmount must be a positive number"